back/requirements.txt
flask
flask_cors
Pillow
ultralytics